Re-organise folder structure to move (most of) the web project under /web
Fix a problem with the keycloak realm export file (removed the default policy - it uses a JS type which can't be imported).
