--- src/Keycloak.ts
@@ -1,7 +1,0 @@
-import Keycloak from "keycloak-js";
-
-const keycloak = new Keycloak(
-  
-);
-
-export default keycloak;
