+++ src/web/Keycloak.ts
@@ -1,0 +1,5 @@
+import Keycloak from 'keycloak-js';
+
+const keycloak = new Keycloak();
+
+export default keycloak;
