+++ src/web/components/Core/SnailTrail.scss
