+++ src/web/screens/login.scss
@@ -1,5 +1,3 @@
-@import '/src/colors';
-
 .uid2-login {
   display: grid;
   flex-direction: column;
